feat: add modal error and api error handling

+++ frontend/LAB_TGL_MOBILE/src/shared/services/api/axios.config.ts
@@ -1,8 +1,9 @@
 import axios, { AxiosError } from 'axios';
+import { mountError } from '@shared/util';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 const api = axios.create({
-    baseURL: 'http://192.168.1.22:3333',
+    baseURL: 'http://192.168.18.10:3333',
     headers: { 'Content-Type': 'application/json' }
 });
 
@@ -20,25 +21,32 @@ api.interceptors.request.use(
     },
 
     (error: AxiosError) => {
-        return Promise.reject(error);
+        const errorFormated = mountError(error);
+        return Promise.reject(errorFormated);
     }
 );
 
 api.interceptors.response.use(
     async (response) => {
         
-        const token = response.data.token;
+        const urlRequest = response.request._url;
 
-        if( token ) {
-            await AsyncStorage.clear();
-            await AsyncStorage.setItem('@token', JSON.stringify(token));
-        }
+        // Verificar se o token não é do reset-password
+        if( urlRequest.match(/.{1,}\/reset+$/g) ) {
+            const token = response.data.token;
     
+            if( token ) {
+                await AsyncStorage.clear();
+                await AsyncStorage.setItem('@token', JSON.stringify(token));
+            }
+        }
+        
         return response;
     },
 
     (error: AxiosError) => {
-        return Promise.reject(error);
+        const errorFormated = mountError(error);
+        return Promise.reject(errorFormated);
     }
 );
